feat: add dialog opening functions and highlight text component
- add open functions for add-comment, add-post, update-comment, and update-post dialogs
- create HighlightText component for text highlighting functionality
- update existing dialog components to utilize new open functions
- enhance post-management-page with new hooks for managing posts and comments
- refactor UI components for better organization and readability
