+++ .fsdrc
@@ -1,24 +1,12 @@
 {
   "root": "src",
   "layers": {
-    "shared": {
-      "order": 1
-    },
-    "entities": {
-      "order": 2
-    },
-    "features": {
-      "order": 3
-    },
-    "widgets": {
-      "order": 4
-    },
-    "views": {
-      "order": 5
-    },
-    "core": {
-      "order": 6
-    }
+    "core": { "order": 1 },
+    "shared": { "order": 2 },
+    "entities": { "order": 3 },
+    "features": { "order": 4 },
+    "widgets": { "order": 5 },
+    "views": { "order": 6 }
   },
   "segments": ["ui", "model", "api", "lib", "libs", "hooks", "assets"]
 }
