+++ src/features/add-comment/ui/index.ts
@@ -1,1 +1,2 @@
 export * from "./add-comment-dialog"
+export * from "./open"
