+++ src/features/add-post/ui/index.ts
@@ -1,1 +1,2 @@
 export * from "./add-post-dialog"
+export * from "./open"
