+++ src/features/update-comment/ui/index.ts
@@ -1,1 +1,2 @@
 export * from "./update-comment-dialog"
+export * from "./open"
