+++ src/features/update-post/ui/index.ts
@@ -1,1 +1,2 @@
 export * from "./update-post-dialog"
+export * from "./open"
