+++ src/shared/ui/highlight-text/index.ts
@@ -1,0 +1,1 @@
+export * from "./highlight-text"
