+++ src/shared/ui/index.ts
@@ -5,3 +5,4 @@ export * from "./input"
 export * from "./select"
 export * from "./table"
 export * from "./textarea"
+export * from "./highlight-text"
