+++ src/views/post-management-page/view.tsx
@@ -1,13 +1,5 @@
-import { useQueryParamsPagination } from "@/shared/hooks"
 import { Button, Card, CardContent, CardHeader, CardTitle } from "@/shared/ui"
-import { deleteComment as deleteCommentAction, likeComment as likeCommentAction } from "@/entities/comments"
 import type { Post } from "@/entities/posts"
-import {
-  addPost as addPostAction,
-  optimisticAddPost,
-  postEntityQueries,
-  updatePost as updatePostAction,
-} from "@/entities/posts"
 import type { User } from "@/entities/users"
 import { openAddPostDialog } from "@/features/add-post"
 import { openUpdatePostDialog } from "@/features/update-post/ui"
@@ -15,43 +7,12 @@ import { openPostDetailDialog } from "@/widgets/post-detail-dialog"
 import { PostListTable, PostListTableFilter, PostListTablePagination } from "@/widgets/post-list-table"
 import { openUserInfoDialog } from "@/widgets/user-info-dialog"
 
-import { useMutation, useQueryClient } from "@tanstack/react-query"
+import { usePostManagement } from "./model"
+
 import { Plus } from "lucide-react"
 
 const PostManagementView = () => {
-  const queryClient = useQueryClient()
-
-  const [queryParamsPagination] = useQueryParamsPagination()
-
-  const addPost = useMutation({
-    mutationFn: addPostAction,
-    onError: (error) => {
-      console.error("게시물 추가 오류:", error)
-    },
-    onSuccess: (addPostResponse) => {
-      queryClient.setQueryData(
-        postEntityQueries.getPosts({ limit: queryParamsPagination.limit, skip: queryParamsPagination.skip }).queryKey,
-        (prevPostResponse) => optimisticAddPost(prevPostResponse, addPostResponse),
-      )
-    },
-  })
-
-  const updatePost = useMutation({
-    mutationFn: updatePostAction,
-    onError: (error) => {
-      console.error("게시물 업데이트 오류:", error)
-    },
-  })
-
-  const deleteComment = useMutation({
-    mutationFn: deleteCommentAction,
-    onError: (error) => console.error("댓글 삭제 오류:", error),
-  })
-
-  const likeComment = useMutation({
-    mutationFn: likeCommentAction,
-    onError: (error) => console.error("댓글 좋아요 오류:", error),
-  })
+  const { queryParamsPagination, addPost, updatePost, deleteComment, likeComment } = usePostManagement()
 
   const handleOpenPostDetailDialog = (post: Post) => {
     openPostDetailDialog({
