+++ src/widgets/user-info-dialog/index.ts
@@ -1,1 +1,2 @@
-export { UserInfoDialog, openUserInfoDialog } from "./user-info-dialog"
+export { UserInfoDialog } from "./user-info-dialog"
+export { openUserInfoDialog } from "./open"
