+++ vitest.config.ts
@@ -7,5 +7,7 @@ export default defineConfig({
   test: {
     globals: true,
     environment: "jsdom",
+    setupFiles: ["./vitest.setup.ts"],
+    include: ["**/*.{test,spec}.?(c|m)[jt]s?(x)"],
   },
 })
